pyproject.toml
[project]
name = "autumn-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datetime>=5.5",
    "pathlib>=1.0.1",
    "pygit2>=1.19.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
]
